requirements.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
Django==5.1.2
django-allauth==65.3.0
django-sslserver==0.22
django-widget-tweaks==1.5.0
eventkit==1.0.3
executing==2.1.0
finnhub-python==2.4.20
frozendict==2.4.6
holidays==0.60
html5lib==1.1
ibapi==9.81.1.post1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.7
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.54
